cloudinary changes

+++ src/utils/cloudinary.js
@@ -10,7 +10,6 @@ cloudinary.config({
 
 // http://res.cloudinary.com/dtzyaxndt/image/upload/v1712471612/65d0b8cbcf65c91cd0e0dc07/postImages/g5mtswr6fyb3ea7qcy5s.jpg
 
-
 const uploadOnCloudinary = async (localFilePath, folder) => {
   try {
     if (!localFilePath) return null;
@@ -29,8 +28,39 @@ const uploadOnCloudinary = async (localFilePath, folder) => {
     //upload the file on cloudinary
     const response = await cloudinary.uploader.upload(localFilePath, {
       folder: folder,
+      use_filename: true,
       resource_type: "auto",
-      use_filename:true
+    });
+    // file has been uploaded successfull
+    fs.unlinkSync(localFilePath);
+    return response;
+  } catch (error) {
+    fs.unlinkSync(localFilePath);
+    throw new ApiError(400, error.message);
+  }
+};
+const uploadVideoCloudinary = async (localFilePath, folder,aspectRatio) => {
+  try {
+    if (!localFilePath) return null;
+    // Validate file type (only allow images)
+    // const allowedExtensions = ["jpg", "jpeg", "png", "gif", "webp"]; // Add more if needed
+    // const fileExtension = localFilePath.split(".").pop().toLowerCase();
+    // if (!allowedExtensions.includes(fileExtension)) {
+    //   console.log(
+    //     "Invalid file type. Only images (jpg, jpeg, png, gif, webp) are allowed."
+    //   );
+    //   throw new ApiError(
+    //     404,
+    //     "Invalid file type. Only images (jpg, jpeg, png, gif, webp) are allowed."
+    //   );
+    // }
+    //upload the file on cloudinary
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      folder: folder,
+      use_filename: true,
+      resource_type: "video",
+      transformation: [{ aspect_ratio: aspectRatio, crop: "crop" }],
+      eager: [{ format: "mp4" }],
     });
     // file has been uploaded successfull
     fs.unlinkSync(localFilePath);
@@ -43,13 +73,18 @@ const uploadOnCloudinary = async (localFilePath, folder) => {
 
 const uploadMultipleOnCloudinary = async (localFilePaths = [], folder) => {
   try {
-    const uploadPromises = localFilePaths.map(async (path) => {
-      const res = await uploadOnCloudinary(path, folder);
-      console.log(res)
-      return {
-        url: res.secure_url,
-        publicId: res.public_id
+    const uploadPromises = localFilePaths.map(async (f) => {
+      let res;
+      if (f.isVideo) {
+        res = await uploadVideoCloudinary(f.path, folder, f.aspectRatio);
+      } else {
+        res = await uploadOnCloudinary(f.path, folder);
       }
+      return {
+        url: res?.secure_url,
+        publicId: res?.public_id,
+        type: f.isVideo ? "VIDEO":"IMAGE"
+      };
     });
 
     const uploadedUrls = await Promise.all(uploadPromises);
@@ -64,7 +99,7 @@ const deleteFromCloudinary = async (publicIds = []) => {
     const deletePromises = publicIds.map(async (publicId) => {
       // Delete the file from Cloudinary
       const deletionResponse = await cloudinary.uploader.destroy(publicId);
-      return deletionResponse.result === 'ok';
+      return deletionResponse.result === "ok";
     });
 
     const deletionResults = await Promise.all(deletePromises);
@@ -77,10 +112,18 @@ const deleteFromCloudinary = async (publicIds = []) => {
 const deleteMultipleFromCloudinary = async (publicIds = []) => {
   try {
     const deletionResults = await deleteFromCloudinary(publicIds);
-    return deletionResults.filter(result => result === true);
+    return deletionResults.filter((result) => result === true);
   } catch (error) {
-    throw new Error(`Failed to delete multiple files from Cloudinary: ${error.message}`);
+    throw new Error(
+      `Failed to delete multiple files from Cloudinary: ${error.message}`
+    );
   }
 };
 
-export { uploadOnCloudinary, uploadMultipleOnCloudinary, deleteFromCloudinary, deleteMultipleFromCloudinary  };
+export {
+  uploadOnCloudinary,
+  uploadMultipleOnCloudinary,
+  deleteFromCloudinary,
+  deleteMultipleFromCloudinary,
+  uploadVideoCloudinary,
+};
